requirements.txt
openai-agents
openai
pydantic
asyncio
gradio==6.10.0
python-dotenv
resend

